this code is a dead end

+++ src/audio_library.c
@@ -68,7 +68,7 @@ double audio_library_get(AudioLibrary *audio_library, int i, double t) {
     }
     AudioData *data = &audio_library->audio[i];
     int sample = (int)(t * (double)data->spec.freq);
-    if (sample >= data->length_samples) {
+    if (sample >= (int)data->length_samples) {
         return 0;
     }
     if (SDL_AUDIO_ISFLOAT(data->spec.format)) {
